Adding integration id

+++ database/migrations/2025_12_10_000000_create_companies_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->date('cert_expires_at');
 
             $table->tinyInteger('environment')->default(NfseEnvironment::HOMOLOGATION);
-            $table->unsignedInteger('last_dps_number')->default(0);
+            $table->unsignedBigInteger('last_dps_number')->default(0);
             $table->string('dps_serie', 5)->default('00001');
             
             $table->timestamps();
